+++ src/interface_biosphere_pmodel.mod.f90
@@ -26,9 +26,8 @@ module md_interface_pmodel
     real :: rd_to_vcmax
     real :: tau_acclim
     real :: kc_jmax
-    real :: kphio_par_c
-    real :: kphio_par_d
-    real :: kphio_par_e
+    real :: coldacclim_par_a
+    real :: coldacclim_par_b
     real :: coldacclim_par_c
     real :: coldacclim_par_d
   end type paramstype_calib